requirements.txt
# PYrte Radio Shack Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.23.0
python-socketio>=5.11.0
requests>=2.31.0
slowapi>=0.1.9  # API rate limiting
apscheduler>=3.10.0  # Background task scheduling

# Database (PostgreSQL)
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Redis
redis>=5.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Email
aiosmtplib>=3.0.0

# Audio Processing
pydub>=0.25.0
mutagen>=1.47.0
# udio-wrapper>=0.2.0  # Optional: Install manually if using Udio provider

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
aiosqlite>=0.19.0

# Code Quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# Type Stubs
types-redis>=4.6.0
types-python-dateutil>=2.8.0
